% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{acf_ci}
\alias{acf_ci}
\title{Confidence interval for autocorrelation function

Calculate a permutation-based confidence interval for the
by-subject autocorrelation function}
\usage{
acf_ci(data, subject, residuals, lag.max = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A dataset.}

\item{subject}{Unquoted name of the variable identifying individual
subjects.}

\item{residuals}{Unquoted variable name for residuals.}

\item{lag.max}{Maximum lag at which to calculate the autocorrelation function. See \code{\link{acf_subj}} for default behavior.}

\item{alpha}{Alpha level.}
}
\value{
Table with (1-alpha)% confidence interval for the autocorrelation function at each time lag.
}
\details{
Permutes the residuals for each participant before calculating the confidence interval for the correlation at each time lag.
}
