\name{autocorr-package}
\alias{autocorr-package}
\alias{autocorr}
\docType{package}
\title{
Asymptotic methods and resampling techniques for testing autocorrelations
}
\description{
Two asymptotic methods (the 1/T approximation method and the Bartlett's formula method) and two resampling techniques (the surrogate data method and the vectorized moving block bootstrap) are implemented to test the significance of autocorrelations at certain lags.
}
\details{
\tabular{ll}{
Package: \tab autocorr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-25\cr
License: \tab GPL \cr
}
Different methods for testing autocorrelation are implemented. Two asymptotic methods including the 1/T approximation method and the Bartlett's formula method are implemented. Two resampling techniques including the moving block bootstrap and the surrogate data method are also provided.
}
\author{
Zijun Ke and Zhiyong Zhang
}
\references{
Bartlett, M. S. (1955). An introduction to stochastic processes. Cambridge: Cambridge University Press. (p.289)

Efron, B., & Tibshirani, R. J. (1994). An introduction to the bootstrap. Chapman and Hall/CRC.

Ke, Z. & Zhang, Z. (2015). Testing Autocorrelations: Asymptotic Methods versus Resampling Techniques

Kunsch, H. (1989). The jackknife and the bootstrap for general stationary observations. The Annals of Statistics, 17(3), 1217-1241.

Shumway, R. H., & Stoffer, D. S. (2006). Time series analysis and its applications with r examples (2nd ed.). New York: Springer.  (p.519)

Theiler, J., Eubank, S., Longtin, A., Galdrikan, B., & Farmer, J. D. (1992). Testing for nonlinearity in time series: The method of surrogate data. Physica D, 58, 77-94.

Zhang, G., & Browne, M. W. (2010). Bootstrap standard error estimates in dynamic factor analysis. Multivariate Behavioral Research, 45(453-482).
}
\keyword{ package }
\examples{
## to be added
}
