% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{blst_quantiles}
\alias{blst_quantiles}
\title{Compute quantiles of random effect variances}
\usage{
blst_quantiles(q1 = 0.2, q2 = 0.8)
}
\arguments{
\item{q1}{Lower quantile.}

\item{q2}{Upper quantile.}
}
\value{
A matrix with the quantiles.
}
\description{
Compute quantiles of the random effect variances for the studies
examined by Barr, Levy, Scheepers, and Tily (2013).
}
\examples{
blst_quantiles(.1, .9) # 10\% and 90\% cutoffs

blst_quantiles() # 20\% and 80\% cutoffs

}
\seealso{
\code{\link{blst_studies}}
}
