% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blst_studies}
\alias{blst_studies}
\title{Sample of study random effect variances}
\format{
A tibble with 13 rows and 6 columns:
\describe{
  \item{ID}{Numeric ID of the study (see \url{http://talklab.psy.gla.ac.uk/simgen/realdata.html}).}
  \item{Residual}{Estimate of the residual variance.}
  \item{subj_int}{By-subject random intercept variance as a proportion of the residual variance.}
  \item{subj_slp}{By-subject random slope variance as a proportion of the residual variance.}
  \item{item_int}{By-item random intercept variance as a proporition of the residual variance.}
  \item{item_slp}{By-item random slope variance as a proporition of the residual variance.}
}
}
\usage{
blst_studies
}
\description{
Sample of study random effect variances
}
\details{
A small sample of estimated random effect variances from
  13 studies, first presented in the supplementary materials to the
  paper by
  \href{http://talklab.psy.gla.ac.uk/simgen/realdata.html}{Barr,
  Levy, Scheepers, and Tily (2013).}
}
\keyword{datasets}
