% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{consolidate_results}
\alias{consolidate_results}
\title{Consolidate Simulation Results}
\usage{
consolidate_results(inpath, outpath)
}
\arguments{
\item{inpath}{Path to directory containing simulation results.}

\item{outpath}{Path to directory where results files will be stored.}
}
\description{
Consolidate results from multiple simulations into a smaller set of
files, removing date/time, host, and PID information.
}
