% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gp.R
\name{cov_fn}
\alias{cov_fn}
\title{Generate autocorrelation function}
\usage{
cov_fn(t, sigma = 4, gamma = 3)
}
\arguments{
\item{t}{Vector of time lags.}

\item{sigma}{Standard deviation for the Gaussian process.}

\item{gamma}{Parameter co-determining (with sigma) the length of the autocorrelation.}
}
\value{
Vector of covariances. The duration of the autocorrelation across lags is determined by the ratio of sigma and gamma.
}
\description{
Generate autocorrelation function
}
\details{
Autocorrelation function for a Gaussian process, defined as \code{sigma^2 * exp(-t^2 / (2 * gamma^2))}. To convert covariances to correlations, use \code{cov_fn(t, sigma, gamma) / sigma^2}.
}
\examples{
sig <- 4
plot(cov_fn(1:20, sigma = sig) / sig^2,
     type = 'b', ylim = c(0, 1), ylab = 'correlation', xlab = 'lag')
}
