% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{dw_subj}
\alias{dw_subj}
\title{Durbin-Watson statistic by subject

Calculates the Durbin-Watson statistic by subject.}
\usage{
dw_subj(data, subject, residual)
}
\arguments{
\item{data}{A dataset.}

\item{subject}{Unquoted variable name in \code{data} identifying
individual subjects.}

\item{residuals}{Unquoted variable name in \code{data} of
residuals.}
}
