% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{elogit_test}
\alias{elogit_test}
\title{Fit GLMM to logit with empirical logit as DV}
\usage{
elogit_test(dat)
}
\arguments{
\item{dat}{Data frame generated by the function \code{\link{gen_logit_data}()}.}
}
\value{
p value.
}
\description{
Fit GLMM to logit with empirical logit as DV
}
