% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin.R
\name{fit_2x2_sin}
\alias{fit_2x2_sin}
\title{Fit models to 2x2 data with sine wave errors}
\usage{
fit_2x2_sin(dat, cs = FALSE)
}
\arguments{
\item{dat}{Data generated using \link{sim_2x2_sin}.}

\item{cs}{Fit gamm smooth as main effect of trial (in addition to by-subject factor smooths).}
}
\value{
Array with parameter estimates and standard errors.
}
\description{
Fit models to 2x2 data with sine wave errors
}
