% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logit_jl}
\alias{fit_logit_jl}
\title{Fit logit model via Julia MixedModels package}
\usage{
fit_logit_jl(formula, data, jsess, name = "fm1")
}
\arguments{
\item{jsess}{Julia session (established by \code{JuliaCall::julia_setup()}).}

\item{name}{Name of the Julia object to create.}

\item{mform}{Model formula.}

\item{dat}{Data frame.}
}
\value{
A Julia fitted model object.
}
\description{
Fits a mixed-effects logistic regression model via the Julia MixedModels
package.
}
