% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logit_lmer}
\alias{fit_logit_lmer}
\title{Fit logistic regression model using lmer}
\usage{
fit_logit_lmer(formula, data)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame containing variables referenced in formula.}
}
\value{
If estimation process converges, a \code{glmerMod} object; otherwise \code{NULL}.
}
\description{
Convenience function that fits a logit model and returns NULL when the estimation process fails to converge.
}
