% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{fit_practice_gamm}
\alias{fit_practice_gamm}
\title{Fit Generalized Additive Mixed Model to Practice Data}
\usage{
fit_practice_gamm(dat)
}
\arguments{
\item{dat}{Data frame, the result of \code{\link{sim_practice}}.}
}
\description{
Fit a Generalized Additive Mixed Model (GAMM) to simulated data
containing a practice effect, as generated by
\code{\link{sim_practice}}.
}
\details{
The model is fit using \code{\link[mgcv]{bam}}. It assumes
a constant learning rate across subjects.
}
