% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{fit_practice_lmem}
\alias{fit_practice_lmem}
\title{Fit Linear Mixed Effects Model to Practice Data}
\usage{
fit_practice_lmem(dat)
}
\arguments{
\item{dat}{Data frame, the result of \code{\link{sim_practice}}.}
}
\description{
Fit a Linear Mixed-Effects Model (LMEM) to simulated data
containing a practice effect, as generated by
\code{\link{sim_practice}}.
}
\details{
The model is fit using \code{\link[mgcv]{bam}}, but does
not include any wiggly terms. It is therefore equivalent to a
linear mixed-effects model. It assumes a constant learning rate
across subjects.
}
