% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{fit_practice_models}
\alias{fit_practice_models}
\alias{fit_practice_nlme}
\alias{fit_practice_gamm}
\alias{fit_practice_lmem}
\title{Fit Models to Simulated Practice Data}
\usage{
fit_practice_nlme(dat, start_vals = c(10, 400, 24, 400, -2))

fit_practice_gamm(dat)

fit_practice_lmem(dat)
}
\arguments{
\item{dat}{Data frame, the result of \code{\link{sim_practice}}.}

\item{start}{Vector of five starting values for
\code{\link[nlme]{nlme}}, for the values (1) within-effect; (2)
asymptote; (3) between-effect; (4) sweep; and (5) learning rate.}
}
\description{
Fit Models to Simulated Practice Data
}
\details{
\code{fit_practice_nlme} fits a Non-Linear Mixed-Effects
(NLME) model using \code{\link[nlme]{nlme}};
\code{fit_practice_gamm} fits a Generalized Additive
Mixed-Effects Model (GAMM) using \code{\link[mgcv]{bam}} with a
by-subject factor smooth for trial; \code{fit_practice_lmem} fits
a Linear Mixed-Effects Model (LMEM) also using
\code{\link[mgcv]{bam}} for comparability to the GAMM, but as it
includes no wiggly terms, it is functionally equivalent to a
model fit using the \code{lme4} package. For simplicity, all
models assume a constant learning rate across subjects.

The exact models are:


}
\section{NLME}{


\code{nlme::nlme(y ~ b0 * wij + Asym + sweep * exp(-exp(lrc) * tij),
                     groups = ~ subj_id,
                     dat, fixed = list(b0 ~ 1, Asym ~ bi, sweep ~ 1, lrc ~ 1),
                     random = nlme::pdDiag(b0 + Asym + sweep ~ 1),
                     start = start_vals)}
}

\section{LMEM}{


\code{mgcv::bam(y ~ wij + bi +
             s(subj_id, bs = "re") + # random intercept
             s(wij, subj_id, bs = "re"), # random slope
           data = dat)}
}

\section{GAMM}{


\code{mgcv::bam(y ~ wij + bi +
             s(tij, bs = "tp") + # common smooth
             s(wij, subj_id, bs = "re") + # random slope
             s(tij, subj_id, bs = "fs"), # factor smooth (time-varying icept)
           data = dat)}
}

