% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{fit_practice_nls}
\alias{fit_practice_nls}
\title{Fit Non-linear Mixed-Effects Model to Practice Data}
\usage{
fit_practice_nls(dat, start_vals = c(30, 400, 200, 400, -2))
}
\arguments{
\item{dat}{Data frame, the result of \code{\link{sim_practice}}.}

\item{start}{Vector of five starting values for
\code{\link[nlme]{nlme}}, for the values (1) within-effect; (2)
asymptote; (3) between-effect; (4) sweep; and (5) learning rate.}
}
\description{
Fit a non-linear mixed-effects ('nlme') model to simulated data
containing a practice effect, as generated by
\code{\link{sim_practice}}.
}
\details{
This model assumes a constant learning rate across subjects.
}
