% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroop_simulate.R
\name{fit_stroop}
\alias{fit_stroop}
\title{Fit GAMM and LMEM Models to Simulated Stroop Data}
\usage{
fit_stroop(dat)
}
\arguments{
\item{dat}{A data.frame with simulated stroop data, the result of
\code{\link{simulate_stroop}}.}
}
\value{
A named vector with the following statistical results:

\describe{

\item{\code{G.e.int}}{Estimated fixed intercept of the GAMM model.}

\item{\code{G.e.A}}{Estimated fixed effect of A of the GAMM model.}

\item{\code{G.e.B}}{Estimated fixed effect of B of the GAMM model.}

\item{\code{G.se.int}}{Standard error for the intercept of the GAMM model.}

\item{\code{G.se.A}}{Standard error for effect A of the GAMM model.}

\item{\code{G.se.B}}{Standard error for effect B of the GAMM model.}

\item{\code{G.p.int}}{P-value for the intercept of the GAMM model.}

\item{\code{G.p.A}}{P-value for A of the GAMM model.}

\item{\code{G.p.B}}{P-value for B of the GAMM model.}

\item{\code{L.e.int}}{Estimated fixed intercept of the LMEM model.}

\item{\code{L.e.A}}{Estimated fixed effect of A of the LMEM model.}

\item{\code{L.e.B}}{Estimated fixed effect of B of the LMEM model.}

\item{\code{L.se.int}}{Standard error for the intercept of the LMEM model.}

\item{\code{L.se.A}}{Standard error for effect A of the LMEM model.}

\item{\code{L.se.B}}{Standard error for effect B of the LMEM model.}

\item{\code{L.p.int}}{P-value for the intercept of the LMEM model.}

\item{\code{L.p.A}}{P-value for A of the LMEM model.}

\item{\code{L.p.B}}{P-value for B of the LMEM model.}

}
}
\description{
Uses \code{\link[mgcv]{bam}} to fit a Generalized Additive
Mixed-Effects Model (GAMM) and a Linear Mixed-Effects Model (LMEM)
to simulated Stroop data.
}
\details{
The models fit to the data are:
}
\section{GAMM}{

\code{bam(latency ~ A_c + B_c +
      s(trial, bs = "tp") +                    # common smooth
      s(session_id, trial, bs = "fs", m = 1) + # factor smooth
      s(A_c, session_id, bs = "re"),           # random slope
      data = dat)}
}

\section{LMEM}{

\code{bam(latency ~ A_c + B_c +
      s(session_id, bs = "re") +  # random intercept
      s(A_c, session_id, bs = "re"), # random slope
      data = dat)}
}

