% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gamm_2x2}
\alias{gamm_2x2}
\title{Simulate and Fit GAMM}
\usage{
gamm_2x2(n_subj, n_obs, fixed, amx, amx_wt = NULL)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject.}

\item{fixed}{vector of fixed effects in the following order:
Intercept, main effect of A, main effect of B, and AB
interaction.}

\item{amx}{Matrix of autocorrelation functions.}

\item{amx_wt}{Weights for sampling rows from \code{amx}.}
}
\description{
Simulate 2x2 data and fit GAMM models.
}
