% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gen_logit_data}
\alias{gen_logit_data}
\title{Generate simulated nested categorical time-series data}
\usage{
gen_logit_data(n_subj, n_obs, fixed, downsample = 1L)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject (multiple of 2).}

\item{fixed}{Fixed effects parameters (1st element = intercept, 2nd element = main effect).}

\item{downsample}{How much to downsample from original 1000 Hz data. (1 = 1000Hz, 2 = 500Hz, 4 = 250Hz, etc.)}
}
\value{
Data frame with time series data.
}
\description{
Generate simulated nested categorical time-series data
}
