% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm_2x2.R
\name{kkl_grafted}
\alias{kkl_grafted}
\title{Simulate and Fit GAMM}
\usage{
kkl_grafted(n_subj, n_obs, fixed, is_acf, amr, amr_wt = NULL,
  replace = TRUE)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject.}

\item{fixed}{vector of fixed effects in the following order:
Intercept, main effect of A, main effect of B, and AB
interaction.}

\item{is_acf}{Is argument \code{amr} a list of residuals (TRUE) or a matrix
of autocorrelation functions (FALSE)?}

\item{amr}{List of residuals or matrix of autocorrelation functions, with each element or row representing a different subject.}

\item{amr_wt}{Weights for sampling rows from \code{amx}.}

\item{replace}{Whether to sample residual vectors with replacement.}
}
\description{
Simulate 2x2 data, graft KKL residuals, estimate models
}
