% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kkl_mx}
\alias{kkl_mx}
\title{Residuals from the KKL dataset (matrix format)}
\format{
A matrix with 86 rows and 800 columns. Each row is an individual subject, each column a trial in sequential order.
}
\source{
The raw KKL data can be found at \url{https://github.com/dmbates/RePsychLing}
}
\usage{
kkl_mx
}
\description{
Residuals from the KKL dataset (matrix format)
}
\details{
Residuals from the KKL dataset in matrix format, after fitting a linear-mixed effects model with random intercepts and random slopes. They represent the naturally occurring autocorrelation patterns for the 86 participants in the KKL dataset, which is available through the \code{RePsychLing} package. The script for fitting the model and extracting the residuals is available at \url{https://github.com/dalejbarr/autocorr/data-raw/extract_kkl_resids.R}.
}
\keyword{datasets}
