% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversampled.R
\name{oversampled_ts}
\alias{oversampled_ts}
\title{Generate single oversampled categorical time series}
\usage{
oversampled_ts(p_targ, prob = 0.005, downsample = 1L,
  sacc_overhead = 200L)
}
\arguments{
\item{p_targ}{Probability of gazing at the target.}

\item{prob}{Probability of a decision at each frame; 1 / prob gives mean lag between decisions.}

\item{downsample}{1L for 1000 Hz; 2L for 500 Hz; 4L for 250 Hz, etc.}

\item{sacc_overhead}{Saccadic overhead.}
}
\value{
Oversampled time series, represented by a vector of zeroes and ones
}
\description{
Generate single oversampled categorical time series
}
\details{
Decision lags are sampled from a geometric distribution with probability \code{prob}, with each decision offset from the last by \code{sacc_overhead}.
}
