% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/permute.R
\name{permutation_test}
\alias{permutation_test}
\alias{permutation_test}
\title{Run a permutation test on logit data}
\usage{
permutation_test(dat, nmc = 1000L)

permutation_test(dat, nmc = 1000L)
}
\arguments{
\item{dat}{Data generated by a call to \code{\link{gen_logit_data}()}.}

\item{nmc}{Number of Monte Carlo runs.}

\item{dat}{Data frame generated by the function \code{\link{gen_logit_data}()}.}

\item{nmc}{Number of Monte Carlo runs.}
}
\value{
p-value.

p value.
}
\description{
Run a permutation test on logit data

Run a permutation on logistic regression datasets.
}
\details{
This function is only to be used for data generated by \code{gen_logit_data()}.
}
