% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{permute_resids}
\alias{permute_resids}
\title{Permute residuals}
\usage{
permute_resids(data, subject, residuals)
}
\arguments{
\item{data}{The dataset.}

\item{subject}{Unquoted name of the variable identifying individual
subjects.}

\item{residuals}{Unquoted variable name for residuals.}
}
\value{
A table with the \code{residuals} column permuted for each subject.
}
\description{
Independently permute the residuals for each subject in the data set.
}
