% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resids_to_rarray}
\alias{resids_to_rarray}
\title{Convert residuals to ragged array (list)}
\usage{
resids_to_rarray(data, subject, residuals)
}
\arguments{
\item{data}{The dataset.}

\item{subject}{Unquoted name of the variable identifying individual
subjects.}

\item{residuals}{Unquoted variable name for residuals.}
}
\value{
A list where each element is the vector of residuals for a subject.
}
\description{
Convert a column of residuals in a data frame into a list.
}
