% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_2x2}
\alias{run_2x2}
\title{Simulate and fit 2x2 autocorrelated data}
\usage{
run_2x2(n_subj, n_obs, fixed, jsess, amx, amx_wt = NULL)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject.}

\item{fixed}{vector of fixed effects in the following order:
Intercept, main effect of A, main effect of B, and AB
interaction.}

\item{jsess}{Julia session.}

\item{amx}{Matrix of autocorrelation functions.}

\item{amx_wt}{Weights for sampling rows from \code{amx}.}
}
\value{
A two dimensional array with fixed effects estimates and
  standard errors.
}
\description{
Simulate data from a 2x2 mixed design (A within and B between) with
autocorrelated residuals.
}
\details{
Simulate data corresponding to three models: no
  autocorrelation (\code{mod_no}), random trial order
  (\code{mod_rand}), and with trials blocked by levels of A
  (\code{mod_block}). Models are fit via Julia (using
  \code{\link{run_jl}}).
}
\seealso{
\code{\link{sim_2x2}}, \code{\link{fit_jl}}
}
