% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle Trial Order}
\usage{
shuffle(n_subj, n_obs)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject.}
}
\value{
An unnamed list of length \code{n_subj}, with each element
of that list being a vector of length \code{n_obs} containing the
position number for each trial.
}
\description{
Randomly shuffle the order that trials appear for a set of participants.
}
\details{
This is the default function used to perform the
randomization of within-subject factor labels when generating
simulated data using \code{\link{sim_2x2}}. The original,
unrandomized dataset is ordered by the variables \code{B},
\code{subj_id}, \code{A}, such that the first half of subjects
are in \code{B1} and the second half in \code{B2}, and the first
half of the observations for subject \code{i} are in condition A1
and the last half are in condition \code{A2}. This function
generates indices to use to re-order the trials randomly for each
subject. It is simply a wrapper around the function
\code{replicate(n_subj, sample(seq_len(n_obs)), simplify=FALSE)}.

It is possible to write your own function to create pseudorandom
series. Any user-defined function must have the same formal
arguments as this function in the same order, and return a list of
integer vectors, with each vector of length \code{n_obs}, and the
list itself of length \code{n_subj}.
}
\examples{
shuffle(4, 8)

sim_2x2(4, 8, rand_fn="shuffle") ## the default for sim_2x2

## user-defined randomization function
my_shuffle <- function(n_subj, n_obs) {
  odds <- seq(1, n_obs, 2)
  evens <- seq(2, n_obs, 2)
  replicate(n_subj,
            if (sample(c(FALSE, TRUE), 1)) {
              c(odds, evens)
            } else {
              c(evens, odds)
            },
            simplify=FALSE)
}

my_shuffle(4, 8)
sim_2x2(4, 8, rand_fn="my_shuffle")

}
