% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{shuffle_alt}
\alias{shuffle_alt}
\title{Alternating Trial Order}
\usage{
shuffle_alt(n_subj, n_obs)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_obs}{Number of observations per subject.}
}
\value{
An unnamed list of length \code{n_subj}, with each element
of that list being a vector of length \code{n_obs} containing the
position number for each trial.
}
\description{
Generate trial order so that the levels of the within-subject
factor form an alternating sequence.
}
\examples{
shuffle_alt(4, 4)

sim_2x2(4, 4, rand_fn="shuffle_alt")

}
