% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_2x2_new.R
\name{sim_2x2_crossed}
\alias{sim_2x2_crossed}
\title{Simulate 2x2 data with autocorrelated errors}
\usage{
sim_2x2_crossed(
  n_subj = 48,
  n_items = 48,
  int = 0,
  A = 0,
  B = 0,
  AB = 0,
  item_rint = 0.5,
  item_rslp = 0.5,
  item_rcorr = 0.5,
  subj_rint = 0.5,
  subj_rslp = 0.5,
  subj_rcorr = 0.5,
  version = 0L,
  verbose = FALSE,
  extra_args = NULL
)
}
\arguments{
\item{n_subj}{Number of subjects to simulate. Must be a positive
integer that is a multiple of 2.}

\item{n_items}{Number of items/observations per subject. Must be a positive
even integer.}

\item{int}{Intercept.}

\item{A}{Main effect of A (within-subject factor).}

\item{B}{Main effect of B (between-subject factor).}

\item{AB}{AB interaction effect.}

\item{item_rint}{Item random intercept variance.}

\item{item_rslp}{Item random slope variance (for factor B).}

\item{item_rcorr}{Item random correlation.}

\item{subj_rint}{Subject random intercept variance.}

\item{subj_rslp}{Subject random slope variance (for factor A).}

\item{subj_rcorr}{Subject random correlation.}

\item{version}{How to generate residuals: either an integer
representing the Scenario number (see \code{\link{errsim}}) or
the name of a user-defined function.}

\item{verbose}{Whether the data frame should GLM components.}

\item{extra_args}{A list of extra arguments to be passed to a
user-defined function to generate residuals.}
}
\value{
A data frame with \code{n_subj * n_items} rows and either 9
or 12 columns depending on whether verbose is TRUE or FALSE
respectively.

\describe{

\item{\code{subj_id}}{}

\item{\code{A}}{Level of within-subject factor A.}

\item{\code{B}}{Level of between-subject factor B.}

\item{\code{A_c}}{Deviation-coded predictor for A.}

\item{\code{B_c}}{Deviation-coded predictor for B.}

\item{\code{tnum_r}}{Trial number for the fully randomized version.}

\item{\code{tnum_b}}{Trial number for the blocked version.}

\item{\code{Y_r}}{Response variable for the fully randomized version.}

\item{\code{Y_b}}{Response variable for the fully blocked version.}

\item{\code{item_rint}}{Item random intercept effect (verbose mode only.)}

\item{\code{item_rslp}}{Item random slope effect (verbose mode only).}

\item{\code{subj_rint}}{Subject random intercept effect (verbose mode only.)}

\item{\code{subj_rslp}}{Subject random slope effect (verbose mode only).}

\item{\code{Y_fit}}{Fitted value (all effects except residual.)}

}
}
\description{
Simulate 2x2 data with autocorrelated errors
}
