% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_2x2_kkl}
\alias{sim_2x2_kkl}
\title{Simulate data with naturalistic autocorrelated errors}
\usage{
sim_2x2_kkl(n_subj, fixed, re_range)
}
\arguments{
\item{n_subj}{Number of subjects to simulate.}

\item{fixed}{Vector of population parameters for the fixed effects (intercept, main effect of A, main effect of B, AB interaction).}

\item{re_range}{Two-element vector defining the range for random effect variance (intercept and slope of A).}
}
\value{
A data frame, with:
\describe{
   \item{ts_r}{the trial number given random ordering}
   \item{ts_b}{the trial number given blocked allocation of trials by the within-factor "A"}
   \item{Y_fit}{the 'fitted' value (combining fixed and random effects but not residual error)}
   \item{Y_acr}{the response variable (random order)}
   \item{Y_acb}{the response variable (blocked order)}
}
}
\description{
Simulate data with naturalistic autocorrelated errors
}
\details{
Simulates data from a 2x2 mixed design, with factor A
  within subjects and factor B between subjects, with residuals
  randomly sampled from the KKL dataset (see ?KKL in the
  RePsychLing package).
}
\seealso{
\code{\link{kkl_df}}, \code{\link{kkl_mx}}
}
