% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_acerr}
\alias{sim_acerr}
\title{Simulate autocorrelated residual error based on source function}
\usage{
sim_acerr(source_acf, length.out = 48L, sd = 1, center = TRUE)
}
\arguments{
\item{source_acf}{Source autocorrelation function to use.}

\item{length.out}{Length of the output time series.}

\item{center}{Mean center the time series.}

\item{sigma}{Desired standard deviation of output time series.}
}
\value{
A vector of residuals modeled after \code{source_acf}.
}
\description{
Simulate autocorrelated residual error based on source function
}
