% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{sim_practice}
\alias{sim_practice}
\title{Simulate data with a practice effect}
\usage{
sim_practice(
  n_subj,
  n_trials,
  within_eff = 10,
  between_eff = 24,
  asymptote = 400,
  sweep = 400,
  learning_rate = -2,
  rslope_sd = 20,
  rasym_sd = 40,
  rsweep_sd = 40,
  rlrate_sd = 0,
  err_sd = 20,
  verbose = FALSE
)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_trials}{Number of trials per subject.}

\item{within_eff}{Mean effect of within-subject factor, coded by
\code{xi}.}

\item{between_eff}{Mean effect of between-subject factor, coded by
\code{xj}.}

\item{asymptote}{Asymptotic value.}

\item{sweep}{The 'sweep' of the decline; i.e., the distance from
the asymptote to the starting value.}

\item{learning_rate}{Speed of the decay.}

\item{rslope_sd}{By-subject standard deviation for the random slope.}

\item{rasym_sd}{By-subject standard deviation for the random asymptote.}

\item{rsweep_sd}{By-subject standard deviation for the random sweep.}

\item{rlrate_sd}{By-subject standard deviation for the learning rate.}

\item{err_sd}{Error standard deviation.}

\item{verbose}{Whether to return random effects in the data frame.}
}
\value{
A data frame with some or all of the following fields,
depending on the value of \code{verbose}:

\describe{
\item{\code{subj_id}}{Unique subject identifier.}
\item{\code{tij}}{Trial number.}
\item{\code{wij}}{Deviation-coded within-subject predictor.}
\item{\code{bi}}{Deviation-coded between-subject predictor.}
\item{\code{y}}{The response value.}
\item{\code{S0i}}{By-subject random slope.}
\item{\code{S1i}}{By-subject random asymptote.}
\item{\code{S2i}}{By-subject random sweep.}
\item{\code{S3i}}{By-subject random learning rate.}

\item{\code{trueval}}{Response value excluding the within-subject
effect/random slope.}
}
}
\description{
Simulates psychological response time data containing a practice
effect, such that the response time decays exponentially at a rate
determined by \code{learning_rate} until reaching
\code{asymptote}. Units are milliseconds.
}
\details{
The data contains main effects of the within-subject and
between-subject factors, but no interaction. The model also
contains four by-subject random effects (within-subject slope,
asymptote, sweep, and learning rate). These random effects are
all independent from one another. The data-generating model for
response \code{y} for subject \code{i} and observation {j} is:

\code{yij ~ (g00 + S0i) * wij + g10 + S1i + g11 * bi + 
      (g20 + S2i) * exp(-exp(g30 + S3i) * tij) + eij}

where the predictors are:

\describe{
\item{\code{g00}}{mean within-subject effect;}

\item{\code{S0i}}{by-subject random slope;}

\item{\code{wij}}{deviation-coded predictor for the
within-subject effect;}

\item{\code{g10}}{mean asymptote;}

\item{\code{S1i}}{random asymptote for subject \code{i};}

\item{\code{g11}}{mean between-subject effect;}

\item{\code{bi}}{deviation-coded predictor for the
between-subject effect;}

\item{\code{g20}}{mean sweep (distance from starting value
to asymptote);}

\item{\code{S2i}}{random sweep for subject \code{i};}

\item{\code{g30}}{mean learning rate;}

\item{\code{S3i}}{random learning rate for subject \code{i};}

\item{\code{tij}}{trial number for subject \code{i}, observation \code{j};}

\item{\code{eij}}{random error for subject \code{i}, observation \code{j}.}
}
}
