% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_effects.R
\name{simfit_practice_all}
\alias{simfit_practice_all}
\title{Simulate Practice-Effect Data and Fit All Models}
\usage{
simfit_practice_all(
  n_subj = 48,
  n_trials = 48,
  within_eff = 10,
  between_eff = 24,
  asymptote = 400,
  sweep = 400,
  learning_rate = -2,
  rslope_sd = 20,
  rasym_sd = 40,
  rsweep_sd = 40,
  rlrate_sd = 0,
  err_sd = 20,
  verbose = FALSE
)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_trials}{Number of trials per subject.}

\item{within_eff}{Mean effect of within-subject factor, coded by
\code{xi}.}

\item{between_eff}{Mean effect of between-subject factor, coded by
\code{xj}.}

\item{asymptote}{Asymptotic value.}

\item{sweep}{The 'sweep' of the decline; i.e., the distance from
the asymptote to the starting value.}

\item{learning_rate}{Speed of the decay.}

\item{rslope_sd}{By-subject standard deviation for the random slope.}

\item{rasym_sd}{By-subject standard deviation for the random asymptote.}

\item{rsweep_sd}{By-subject standard deviation for the random sweep.}

\item{rlrate_sd}{By-subject standard deviation for the learning rate.}

\item{err_sd}{Error standard deviation.}

\item{verbose}{Whether to return random effects in the data frame.}
}
\value{
A 44-element numeric vector, containing statistical results
for all three models. Paramters are prefixed by \code{G} for GAMM
results, by \code{L} for LMEM results, and by \code{N} for NLME
results. Following the prefix, is the type of statistic, with
\code{est} for parameter estimates, \code{se} for standard
errors, \code{t} for t-values, and \code{p} for
p-values. Finally, the suffix identifies the parameter in
question, with \code{int} for the intercept, \code{wij} for the
within-subject effect, and \code{bi} for the between-subject
effect. There are also specific parameters for the nonlinear
model, including \code{asym} for the asymptote, \code{sweep} for
the sweep, and \code{lrc} for the learning rate. In cases where
the non-linear model estimation fails, the error is trapped and
the corresponding \code{N.xxx.xxx} values will be set to
\code{NA}.
}
\description{
Simulate data using \code{\link{sim_practice}} and then fit three
models: a non-linear mixed-effects model (NLME), a linear
mixed-effects model (LMEM) and a generalized additive mixed-effects
model (GAMM).
}
\details{
Used in Monte Carlo simulation. See
\code{\link{sim_practice}} for details on data generation and
\code{\link{fit_practice_models}} for details on model fitting.
}
\seealso{
\code{\link{sim_practice}} for information about data
generation, and \code{\link{fit_practice_models}} for details
about model fitting.
}
