% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin.R
\name{sine_sim}
\alias{sine_sim}
\title{Monte Carlo Simulation of Sine Wave Autocorrelation}
\usage{
sine_sim(nmc, nsubj = 48L, ntrials = 48L, fixed = rep(0, 4),
  err_range = c(1, 3), re_range = c(1, 3), varying_phase = TRUE,
  varying_amp = FALSE)
}
\arguments{
\item{nmc}{Number of Monte Carlo runs.}

\item{nsubj}{Number of subjects per dataset.}

\item{ntrials}{Number of trials per subject.}

\item{fixed}{Fixed effect size (four-element vector with intercept, main effect of A, main effect of B, and AB interaction).}

\item{varying_phase}{Does the sine-wave autocorrelation have varying (vs fixed) phase over subjects?}

\item{varying_amp}{Does the sine-wave autocorrelation have varying (vs fixed) amplitude over subjects?}

\item{err}{Error variance (sigma^2).}

\item{re_varmax}{Maximum of the range for the random effect variance parameters.}
}
\value{
An array with labeled dimensions of class 'sinemcs'.
}
\description{
Monte Carlo Simulation of Sine Wave Autocorrelation
}
