% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gp.R
\name{stat_gp}
\alias{stat_gp}
\title{Generate Gaussian time-series with autocorrelation}
\usage{
stat_gp(sigma = 4, gamma = 3, dt_GP = 0.8)
}
\arguments{
\item{sigma}{Standard deviation for the gaussian.}

\item{gamma}{The correlation length.}

\item{dt_GP}{Desired time step for the series.}
}
\value{
TODO
}
\description{
Generate Gaussian time-series with autocorrelation
}
\details{
TODO
}
