% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroop_ML3}
\alias{stroop_ML3}
\title{Data from the Many Labs 3 Stroop Task}
\format{
A tibble with 210,231 observations on 8 variables:

\describe{

\item{\code{session_id}}{Unique session (participant) identifier.}

\item{\code{trial}}{Trial number (0 to 62).}

\item{\code{ink_color}}{Color of the word.}

\item{\code{word}}{Identity of the word.}

\item{\code{congruent}}{Whether the trial was congruent or incongruent.}

\item{\code{cong}}{Deviation-coded predictor for \code{congruent} used in modeling.}

\item{\code{response}}{Color name produced by the participant.}

\item{\code{latency}}{Latency of the response, in milliseconds.}

}
}
\source{
\url{https://osf.io/n8xa7/}
}
\usage{
stroop_ML3
}
\description{
Raw data from the Stroop component of the Many Labs 3 study \insertCite{ML3}{autocorr}.
}
\details{
Unlike the original data, the \code{latency} variable
(originally called \code{trial_latency}) has been set to
\code{NA} for any trials where either (1) an error occurred, or
(2) the latency was greater than 10,000 milliseconds.
}
\examples{
## calculate cell means
with(stroop_ML3, aggregate(latency ~ congruent, stroop_ML3, mean))

}
\references{
\insertAllCited{}
}
\keyword{datasets}
