% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvacf.R
\name{tvacf}
\alias{tvacf}
\title{Time-Varying Autocorrelation Function}
\usage{
tvacf(x, width = 7L)
}
\arguments{
\item{x}{A time series.}
}
\value{
A matrix, the columns vectors of which represent the
results of \code{\link[stats]{acf}} applied to each time slice,
including lags from 1 to \code{width} - 1.
}
\description{
Time-Varying Autocorrelation Function
}
\details{
Slides a moving window along vector \code{x} with
specified \code{width} and calulates the autocorrelation function
at each time slice.
}
\examples{
acseries <- stat_gp(1, 2)$GP[1:50]
tvacf(acseries - mean(acseries), width = 10)
 
}
