% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroop_simulate.R
\name{fit_stroop}
\alias{fit_stroop}
\title{Fit GAMM and LMEM Models to Simulated Stroop Data}
\usage{
fit_stroop(dat)
}
\arguments{
\item{dat}{A data.frame with simulated stroop data, the result of
\code{\link{simulate_stroop}}.}
}
\value{
A named 27-element vector with the following statistical results:

\describe{

\item{\code{Gp.e.B0}}{Estimated fixed intercept of the penalized GAMM model.}

\item{\code{Gp.e.B1}}{Estimated fixed effect of within-subject
factor of the penalized GAMM model.}

\item{\code{Gp.e.B2}}{Estimated fixed effect of between-subject
factor of the penalized GAMM model.}

\item{\code{Gp.se.B0}}{Standard error for the intercept of the penalized GAMM model.}

\item{\code{Gp.se.B1}}{Standard error for within-effect of the penalized GAMM model.}

\item{\code{Gp.se.B2}}{Standard error for between-effect of the penalized GAMM model.}

\item{\code{Gp.p.B0}}{P-value for the intercept of the penalized GAMM model.}

\item{\code{Gp.p.B1}}{P-value for within-effect of the penalized GAMM model.}

\item{\code{Gp.p.B2}}{P-value for between-effect of the penalized GAMM model.}

\item{\code{Gu.e.B0}}{Estimated fixed intercept of the unpenalized GAMM model.}

\item{\code{Gu.e.B1}}{Estimated fixed effect of within-subject
factor of the unpenalized GAMM model.}

\item{\code{Gu.e.B2}}{Estimated fixed effect of between-subject
factor of the unpenalized GAMM model.}

\item{\code{Gu.se.B0}}{Standard error for the intercept of the unpenalized GAMM model.}

\item{\code{Gu.se.B1}}{Standard error for within-effect of the unpenalized GAMM model.}

\item{\code{Gu.se.B2}}{Standard error for between-effect of the unpenalized GAMM model.}

\item{\code{Gu.p.B0}}{P-value for the intercept of the unpenalized GAMM model.}

\item{\code{Gu.p.B1}}{P-value for within-effect of the unpenalized GAMM model.}

\item{\code{Gu.p.B2}}{P-value for between-effect of the unpenalized GAMM model.}

\item{\code{Lm.e.B0}}{Estimated fixed intercept of the LMEM model.}

\item{\code{Lm.e.B1}}{Estimated fixed effect of within-subject
factor of the LMEM model.}

\item{\code{Lm.e.B2}}{Estimated fixed effect of between-subject
factor of the LMEM model.}

\item{\code{Lm.se.B0}}{Standard error for the intercept of the LMEM model.}

\item{\code{Lm.se.B1}}{Standard error for within-effect of the LMEM model.}

\item{\code{Lm.se.B2}}{Standard error for between-effect of the LMEM model.}

\item{\code{Lm.p.B0}}{P-value for the intercept of the LMEM model.}

\item{\code{Lm.p.B1}}{P-value for within-effect of the LMEM model.}

\item{\code{Lm.p.B2}}{P-value for between-effect of the LMEM model.}

}
}
\description{
Uses \code{\link[mgcv]{bam}} to fit two Generalized Additive
Mixed-Effects Models (GAMM) and a Linear Mixed-Effects Model (LMEM)
to simulated Stroop data.
}
\details{
The models fit to the data are:
}
\section{GAMM}{


with penalized factor smooths:

\code{bam(Y_ij ~ W_ij + B_i +
      s(trial, bs = "tp") +                    # common smooth
      s(session_id, trial, bs = "fs", m = 1) + # factor smooth
      s(W_ij, session_id, bs = "re"),           # random slope
      data = dat)}

and unpenalized factor smooths:

\code{bam(Y_ij ~ W_ij + B_i +
      s(trial, bs = "tp") +            # common smooth
      s(session_id, trial, bs = "fs")  # factor smooth
      s(W_ij, session_id, bs = "re"),  # random slope
      data = dat)}
}

\section{LMEM}{


\code{bam(Y_ij ~ W_ij + B_i +
      s(session_id, bs = "re") +      # random intercept
      s(W_ij, session_id, bs = "re"), # random slope
      data = dat)}
}

\examples{
\donttest{
  fit_stroop(simulate_stroop(20))
}

}
