% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mcsim}
\alias{mcsim}
\title{Monte Carlo Simulation of Analysis of Autocorrelated Data}
\usage{
mcsim(
  nmc,
  n_subj = 48L,
  n_obs = 48L,
  A = 0,
  B = 0,
  AB = 0,
  rint_range = blst_quantiles()[, "subj_int"],
  rslp_range = blst_quantiles()[, "subj_slp"],
  rcorr_range = c(-0.8, 0.8),
  version = 0L,
 
    outfile = sprintf("acs_\%05d_\%03d_\%03d_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%d_\%s_\%s_\%s.rds",
    nmc, n_subj, n_obs, A, B, AB, rint_range[1], rint_range[2], rslp_range[1],
    rslp_range[2], version, format(Sys.time(), "\%Y-\%m-\%d-\%H-\%M-\%S"),
    Sys.info()[["nodename"]], Sys.getpid())
)
}
\arguments{
\item{nmc}{Number of Monte Carlo runs.}

\item{n_subj}{Number of subjects to simulate. Must be a positive
integer that is a multiple of 2.}

\item{n_obs}{Number of observations per subject. Must be a positive
even integer.}

\item{A}{Main effect of A (within-subject factor).}

\item{B}{Main effect of B (between-subject factor).}

\item{AB}{AB interaction effect.}

\item{rint_range}{Range of random intercept variance.}

\item{rslp_range}{Range of random slope variance (for factor A).}

\item{rcorr_range}{Range of random correlation.}

\item{version}{Autocorrelation case number; an integer between 0 and 8 (see \code{\link{errsim}}).}

\item{outfile}{Name of output file.}

\item{int}{Intercept.}
}
\value{
Returns NULL.
}
\description{
Monte Carlo Simulation of Analysis of Autocorrelated Data
}
