% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mcsim}
\alias{mcsim}
\title{Monte Carlo Simulation of Data Analysis with Trial-by-Trial Variation}
\usage{
mcsim(
  nmc,
  n_subj = 48L,
  n_obs = 48L,
  A = 0,
  B = 0,
  AB = 0,
  rint_range = blst_quantiles()[, "subj_int"],
  rslp_range = blst_quantiles()[, "subj_slp"],
  rcorr_range = c(-0.8, 0.8),
  version = 0L,
  os_always = is.character(version),
  m = NA,
  k = -1,
  bam_args = NULL,
  fit_blocked = TRUE,
  fit_lmem = TRUE,
 
    outfile = sprintf("acs_\%05d_\%03d_\%03d_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%0.2f_\%s_\%s_\%s_\%s.rds",
    nmc, n_subj, n_obs, A, B, AB, rint_range[1], rint_range[2], rslp_range[1],
    rslp_range[2], if (is.numeric(version)) sprintf("\%02d", version) else version,
    format(Sys.time(), "\%Y-\%m-\%d-\%H-\%M-\%S"), Sys.info()[["nodename"]],
    Sys.getpid()),
  extra_args = NULL,
  rand_fn = "shuffle"
)
}
\arguments{
\item{nmc}{Number of Monte Carlo runs.}

\item{n_subj}{Number of subjects to simulate. Must be a positive
integer that is a multiple of 2.}

\item{n_obs}{Number of observations per subject. Must be a positive
even integer.}

\item{A}{Main effect of A (within-subject factor).}

\item{B}{Main effect of B (between-subject factor).}

\item{AB}{AB interaction effect.}

\item{rint_range}{Range of random intercept variance.}

\item{rslp_range}{Range of random slope variance (for factor A).}

\item{rcorr_range}{Range of random correlation.}

\item{version}{Autocorrelation case number; either an integer
corresponding to Scenario number (see \link{errsim}) or the name
of a user-defined function specified as a character string.}

\item{os_always}{Whether to always fit an overall smooth, or only
for \code{\link[=errsim]{errsim()}} versions 2, 7, 8, 12, 15, and 16. For user-defined
functions, default is to fit a common smooth.}

\item{m}{The \code{m} parameter to be passed on to any factor smooths
(specified in the \code{\link[mgcv:s]{mgcv::s()}} function).}

\item{k}{The \code{k} parameter to be passed on to any factor smooths
(specified in the \code{\link[mgcv:s]{mgcv::s()}} function).}

\item{bam_args}{List of additional arguments to be passed onto
\code{\link[mgcv:bam]{mgcv::bam()}}.}

\item{fit_blocked}{Whether to fit the blocked version in addition
to the randomized version.}

\item{fit_lmem}{Whether to fit the LMEM in addition to the GAMM.}

\item{outfile}{Name of output file.}

\item{extra_args}{Extra args to be passed along to any user-defined
function for generating residuals.}

\item{rand_fn}{Name of the function to randomize trial order
(defaults to \code{\link{shuffle}}). This can be a user-supplied
function; see \code{\link{shuffle}} for details on how this
function should be constructed.}

\item{int}{Intercept.}
}
\value{
Returns NULL.
}
\description{
Monte Carlo Simulation of Data Analysis with Trial-by-Trial Variation
}
\details{
The behavior of \code{os_always} depends on whether
\code{version} has been specified as a scenario number, in which
case it overrides the scenario-dependent selection of an overall
smooth, always fitting an overall smooth. If version is the name
of a user-defined function, then \code{os_always} determines
whether the model includes an overall smooth.
}
