% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permutation_test}
\alias{permutation_test}
\title{Run a permutation test}
\usage{
permutation_test(dat, nmc = 1000L)
}
\arguments{
\item{dat}{Data frame generated by the function \code{\link{gen_logit_data}()}.}

\item{nmc}{Number of Monte Carlo runs.}
}
\value{
p value.
}
\description{
Run a permutation on logistic regression datasets.
}
