% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_2x2}
\alias{sim_2x2}
\title{Simulate 2x2 data with autocorrelated errors}
\usage{
sim_2x2(
  n_subj = 48,
  n_obs = 48,
  int = 0,
  A = 0,
  B = 0,
  AB = 0,
  rint = 0.5,
  rslp = 0.5,
  rcorr = 0.5,
  version = 0L,
  verbose = FALSE
)
}
\arguments{
\item{n_subj}{Number of subjects to simulate. Must be a positive
integer that is a multiple of 2.}

\item{n_obs}{Number of observations per subject. Must be a positive
even integer.}

\item{int}{Intercept.}

\item{A}{Main effect of A (within-subject factor).}

\item{B}{Main effect of B (between-subject factor).}

\item{AB}{AB interaction effect.}

\item{rint}{Random intercept variance.}

\item{rslp}{Random slope variance (for factor A).}

\item{rcorr}{Random correlation.}

\item{version}{Autocorrelation case number; an integer between 0
and 8 (see \code{\link{errsim}}).}

\item{verbose}{Whether the data frame should GLM components.}
}
\value{
A data frame with \code{n_subj * n_obs} rows and either 9
or 12 columns depending on whether verbose is TRUE or FALSE
respectively.

\describe{
\item{\code{subj_id}}{}
\item{\code{A}}{Level of within-subject factor A.}
\item{\code{B}}{Level of between-subject factor B.}
\item{\code{A_c}}{Deviation-coded predictor for A.}
\item{\code{B_c}}{Deviation-coded predictor for B.}
\item{\code{tnum_r}}{Trial number for the fully randomized version.}
\item{\code{tnum_b}}{Trial number for the blocked version.}
\item{\code{Y_r}}{Response variable for the fully randomized version.}
\item{\code{Y_b}}{Response variable for the fully blocked version.}
\item{\code{rint}}{Random intercept effect (verbose mode only.)}
\item{\code{rslp}}{Random slope effect (verbose mode only).}
\item{\code{Y_fit}}{Fitted value (all effects except residual.)}
}
}
\description{
Simulate 2x2 data with autocorrelated errors
}
