% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_2x2}
\alias{sim_2x2}
\title{Simulate data with autocorrelated errors}
\usage{
sim_2x2(n_subj, n_obs, params, is_acf, amx, amx_wt = NULL)
}
\arguments{
\item{n_subj}{Number of subjects to simulate.}

\item{n_obs}{Number of observations per subject.}

\item{params}{Population parameter values for the data generating
process; the structure should be identical to that returned by
\code{\link{funfact::gen_pop}}.}

\item{is_acf}{Is argument \code{amr} a list of residuals (TRUE) or a matrix
of autocorrelation functions (FALSE)?}

\item{amx}{List of residuals or matrix of autocorrelation functions, with each element or row representing a different subject.}

\item{amx_wt}{Weights for sampling elements/rows from \code{amx}; NULL gives equal probability.}
}
\value{
A data frame, with \code{ts_r} as the trial number ordered
  randomly; \code{ts_b} as the trial number blocked by the within
  factor ("A"); \code{Y_fit} as the 'fitted' value (combining fixed
  and random effects but not residual error); \code{Y_acn} as the
  response variable without autocorrelation; \code{Y_acr} as the
  response variable with autocorrelation; and \code{Y_acb} as the
  response variable with autocorrelation for the blocked data.
}
\details{
Simulates data from a 2x2 mixed design, with factor A
  within subjects and factor B between subjects.
}
