% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin.R
\name{sim_2x2_sin}
\alias{sim_2x2_sin}
\title{Simulate data with sine-wave autocorrelated errors}
\usage{
sim_2x2_sin(n_subj, n_obs, fixed, err, phase = TRUE, amp = FALSE,
  re_varmax = 3)
}
\arguments{
\item{n_subj}{Number of subjects to simulate.}

\item{n_obs}{Number of observations per subject.}

\item{fixed}{Vector of population parameters for the fixed effects (intercept, main effect of A, main effect of B, AB interaction).}

\item{err}{Sigma squared, the residual error parameter.}

\item{phase}{Whether the phase of each participant's time-varying function is offset by a random angle (from -pi to pi).}

\item{amp}{Whether the amplitude of each participant's time-varying function is modulated by a random value (from 0 to 2).}

\item{re_varmax}{Maximum variance for random effects parameters (slopes and intercepts).}
}
\value{
A data frame, with \code{ts_r} as the trial number ordered
  randomly; \code{ts_b} as the trial number blocked by the within
  factor ("A"); \code{Y_fit} as the 'fitted' value (combining fixed
  and random effects but not residual error); \code{Y_acr} as the
  response variable with autocorrelation; and \code{Y_acb} as the
  response variable with autocorrelation for the blocked data.
}
\description{
Simulate data with sine-wave autocorrelated errors
}
\details{
Simulates data from a 2x2 mixed design, with factor A
  within subjects and factor B between subjects.
}
