% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroop_simulate.R
\name{simulate_stroop}
\alias{simulate_stroop}
\title{Simulate Stroop Factorial Data}
\usage{
simulate_stroop(
  n_subj,
  A = 0,
  B = 0,
  rivar = diag(autocorr::stroop_mod$covmx)[1],
  rsvar = diag(autocorr::stroop_mod$covmx)[2]
)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{A}{Fixed effect of congruency.}

\item{B}{Fixed effect of the between-subjects factor.}

\item{rivar}{Random intercept variance.}

\item{rsvar}{Random slope variance.}
}
\value{
A data frame with \code{n_subj * 62} simulated observations on
\describe{

\item{\code{session_id}}{A factor with \code{session_id} values
corresponding to the subjects whose residuals were sampled from
\code{\link{stroop_mod[["resid"]]}}}.

\item{\code{trial}}{An integer specifying the trial number (0 to 62).}

\item{\code{A}}{Level of within-subject factor (\code{A1} or \code{A2}).}

\item{\code{B}}{Level of between-subject factor (\code{B1} or \code{B2}).}

\item{\code{A_c}}{Deviation-coded predictor for factor \code{A}.}

\item{\code{B_c}}{Deviation-coded predictor for factor \code{B}.}

\item{\code{latency}}{Simulated response latency in milliseconds.}
}
}
\description{
Simulate data from a mixed factorial design based on the Many Labs
3 (ML3) \insertCite{ML3}{autocorr} Stroop data.
}
\details{
The ML3 data was from a simple design including only the
within-factor of congruency. The simulated data includes an
additional between-subjects factor. The simulation parameters are
taken from \code{\link{stroop_mod}}, with a random subset of
residuals grafted onto the fitted values for each subject.

The simulated response latency for observation \code{j} of subject
\code{i} is given by the following formula:

\code{latency_ij = int + sri_i + B * B_c + (A + srs_i) * A_c + e_ij}

where \code{int} is \code{stroop_mod$fixed["(Intercept)"]},
\code{sri_i} and \code{srs_i} are the random intercept and random
slope for subject \code{i}, \code{A_c} and \code{B_c} are
deviation-coded predictors for the within- and between- subject
factors respectively, \code{A} and \code{B} are the raw within- and
between- subject effect sizes, and \code{e_ij} is the \code{j}th
residual for subject \code{i}.

Note that whereas the original dataset had 63 observations per
subject (with 2/3 in the incongruent condition and 1/3 in the
congruent condition), the current dataset had 1/2 congruent and 1/2
incongruent. To make this work, the 63rd residual was lopped off
for each participant.
}
\examples{
## use parameters from Many Labs 3 model fit
dat <- simulate_stroop(48, A = stroop_mod$fixed["cong"])

mod <- lme4::lmer(latency ~ A_c + B_c + (A_c || session_id), dat)
summary(mod)

}
